infiniband-diags: add rdma-ndd daemon
rdma-ndd is a system daemon which watches the procfs hostname file as well as
monitoring for InfiniBand libudev device changes.

Upon detecting changes it will update the Node Descriptions of the RDMA devices
in the system.

This deamon is intended to work with kernels which support polling of the
procfs hostname.  If your kernel does not support this feature the daemon
will set the Node Descriptions to the hostname at start up only.

ibdiag.conf can be used to specify a node description format with the following
format specifier support.

%h == current hostname
%d == device name

Regardless of previous settings, all NDs are written at startup and
individually when a new device is detected.  Otherwise the ND is compared to
the existing one and not written if it has not changed.

Signed-off-by: Ira Weiny <ira.weiny@intel.com>