libibmad/src/rpc.c: Remove superfluous packet dump
When ibdebug is >= 1, mad_rpc and mad_rpc_rmpp output only
the mad data response in debug output.  It does not output
the request mad data.  This is confusing since it's not clear
if it's request or response output.

When ibdebug > 1, full mad send and receive buffers are
output, which makes the mad data output with ibdebug == 1
superfluous.  Remove this output to remove confusion.

Signed-off-by: Albert Chu <chu11@llnl.gov>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>