infiniband-diags/rdma-ndd: Fix issues with install
On Debian in particular it was reported that the binary was attempting to run
from /sbin when it was installed to /usr/sbin.  This was not seen on RHEL as
RHEL defines /sbin to be a softlink to /usr/sbin.

In addition fix the /var path by using the localstatedir macro.

Reported-by: Dan Ben Yosef <danby@mellanox.com>
Signed-off-by: Ira Weiny <ira.weiny@intel.com>