rsockets: Delay initializing buffers until the inline size is known
The qib HCA ignores the requested max_inline_size on input and
instead returns the supported value on output.  As a result,
the default inline size requested by rsockets is ignored, with 0
being returned.  The code catches this after it creates the QP, but
has already initialized its data buffers prior to creating the QP.

The result is that the inline size used in rs_init_bufs() is larger
than that supported by the qib device.  This causes a failure when
attempting to update available receive buffer space.  The registered
data buffer for the credit message is smaller than what is needed.

Work-around this issue by delaying the initialization of the data
buffers until after the QP has been created and the real size of
the inline data is known.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>