sysfs_file_read: treat a truncate as a failure
Coverity found issue.  When we read a sysfs file and the length of the
read exactly matches our buffer and we don't have a newline to replace
with a null termination, we either have to truncate the result, or fail
to null terminate.  Either way, we will not get the desired behavior, so
in that case, fail the read entirely.

Signed-off-by: Doug Ledford <dledford@redhat.com>