acmp: Specify SGID in PR query
If a PR query is issued with the DGID set, but the SGID not set,
then opensm version 3.3.19 (and likely others) will return a PR
with the SGID set to that of the SM.  This appears to be a bug
with the SM, as an SLID is provided.

In order to be safe, specify the SGID with the PR query, so that
we get a PR that's usable.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>