libcxgb4: skip TERMINATE CQEs before testing for errors
In poll_cq() if the wq happens to have just transitioned into error state,
and a TERMINATE CQE is the next CQE in the hwcq, then the TERMINATE
is incorrectly returned as a successful RECV CQE.  This causes a false
RECV completion.  The fix is to skip TERMINATE CQEs first.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>