Add new poll_cq according to the new CQE format
When working with CQE version 1, the CQE format is different,
therefore we need a new poll_cq function that is compatible with
the new CQE format.

In the coming patches in this series, the library will decide
on the cqe version when allocating the user-context and accordingly
decide which poll_cq to use for this user-context.

Reviewed-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Haggai Abramovsky <hagaya@mellanox.com>
Signed-off-by: Majd Dibbiny <majd@mellanox.com>