Add Raw Packet QP data-path functionality
Raw Ethernet WQE is composed of the following segments:
1. Control segment
2. Eth segment
3. Data segment

The Eth segment contains packet headers and information for stateless
offloading. When posting a Raw Ethernet WQE, the library should copy
the L2 headers into the Eth segment, otherwise the packet is dropped
in the TX. The Data segment should include all the payload except the
headers that were copied to the Eth segment.

Raw Packet QP is composed of RQ and SQ in the hardware. Once the QP's
state is modified to INIT, the RQ is already in RDY state and thus
can receive packets. To avoid this behavior that contradicts the IB
spec, we don't update the doorbell record as long as the QP isn't in
RTR state, and once the QP is modified to RTR, the doorbell record is
update, which allows receiving packets.

Reviewed-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Majd Dibbiny <majd@mellanox.com>