Allow use of huge pages in multiple calls to ibv_fork_init
Setting the environment variable RDMAV_HUGEPAGES_SAFE tells the library
to check the underlying page size used by the kernel for memory regions.
This is required if an application uses huge pages either directly or
indirectly via a library such as libhugetlbfs.

The check of this variable was performed at the first call to
ibv_fork_init. This caused to unpredicted behavior in complex
applications with multiple underlying libraries.

The proposed change will allow support of huge pages without relying on
ibv_fork_init calls order.

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>