Implement memory window bind verb for type 1
In order to bind memory window of type 1, we need to implement
ibv_bind_mw. This is implemented internally by sending a UMR
that fills the data in the MKey and its KLM.

Memory windows of type 1 could be rebind (without being invalidated)
and thus we don't check if the Mkey is FREE or not. Since the
internal implementation uses post_send, but the user isn't allowed
to use this verb in order to bind a type 1 memory window directly,
we split the mlx5_post_send function into an internal function.

Fencing is also implemented here, as the work request after a bind
should be executed with a fence.

Signed-off-by: Matan Barak <matanb@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>