ibcm: Actually use the version script when linking
There is a typo in the Makefile.am that resulted in the version
script being ignored.

This changes the symbol names in the shlib, eg
    50: 0000000000001620   530 FUNC    GLOBAL DEFAULT   11 ib_cm_open_device
becomes
    46: 00000000000015f0   530 FUNC    GLOBAL DEFAULT   12 ib_cm_open_device@@IBCM_1.0

Binaries linked to the old library will continue to work with the new one.

Binaries linked to the new library will print a dynamic linker warning when
using the old library:

  ./a.out: /usr/lib/x86_64-linux-gnu/libibcm.so.1: no version information available (required by ./a.out)

But apparently continue to work.

Drepper (https://www.akkadia.org/drepper/symbol-versioning) seems to indicate
that going forward if we do decide to rev a symbol version then resolution for
binaries linked to the old library will continue to prefer the IBCM_1.0
symbol.  This makes sense as it is the basic mechanism used to introduce
symbol versions in the first place.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>