Replace infiniband/sa-kern-abi.h with the kernel's uapi/rdma/ib_user_sa.h
Using the system header from the kernel is now the expected way to export
definitions to user space. Tree wide update to shift from the local
header and deal with the name changes.

Unfortunately this was exposed as a public installed header, but since
that makes zero sense, and all the names change, just delete it.
Apps are expected to also migrate to rdma/ib_user_sa.h as their
source for this information.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>