libcm: Replace infiniband/cm_abi.h with the kernel's uapi/rdma/ib_user_cm.h
Using the system header from the kernel is now the expected way to export
definitions to user space.

Unfortunately this was exposed as a public installed header, but since
that makes zero sense, and all the names change, just delete it.
Apps are expected to also migrate to rdma/ib_user_cm.h as their
source for this information.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>