Include pthread as part of the link
It is a mistake to not explicitly link to the libraries required.
Not linking causes the symbols to drop the symbol version which could
cause runtime problems down the road if pthread ever goes through
another symbol version change.

Note:
The AC_CHECK_LIB will cause the created Makefile to link with pthread.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>