Avoid gcc 5.4 warning -Wempty-body
gcc 5.4 remarks:
../providers/mlx5/buf.c:95:61: warning: suggest braces around empty body
in an 'if' statement [-Wempty-body]

Around code like this:

        if (shmdt(hmem->shmaddr) == -1)
                mlx5_dbg(stderr, MLX5_DBG_CONTIG, "%s\n",
strerror(errno));

When mlx5_dgb is defined as an empty macro the code expands to
        if (..)
                ;
Which is functionally okay, but strange.

It is much better to make mlx5_dbg an empty static inline, this way it
continues to do parameter type validation even when disabled, and we
can drop few #ifdef MLX5_DEBUG around.

Note:
The compiler is smart enough to drop any assignment that is not
used as of the empty inline function so no penalty is expected in the data
path flows.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>