Improve global COPYING files
Upon deeper analysis it seems there are two variations of the 'OpenIB.org BSD
license', I am calling them the MIT and FreeBSD variants. They differ only in
the text of the disclaimer. The MIT version is used widely as the source
license so that version should be the 'default' in the source tree, not the
FreeBSD version I had selected initially.

Retain the FreeBSD version as we will refer to it in future commits.

Elaborate the COPYING file to discuss the license situation in
more depth.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>