Document copyright situation and select default license
This patch is not intended to change the copyright or license
situation of any of the original code.

Documentation is provided that identifies the various licenses we
have in the source tree.

The default license for new items after the 'Initial commit' merge is
corrected to match the majority license already in use. This is the
license we recommend all new code use. This corrects a mistake I made
in the 'Unified CMake build system' patch which selected the wrong
license file.

Also provide specific guidance on how the 18 different COPYING files
and per-file copyright headers are intended to be interpreted
within the merged tree.

Fixes: 8d4ebd8a9c8 ("Unified CMake build system")
Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>