rxe_cfg: Do not suppress stderr
Suppressing stderr if stdout is already redirected is wrong. Hence
do not suppress stderr. If e.g. the ib_uverbs kernel module is not
loaded, without this patch no error messages are printed. With this
patch applied the following is printed:

Failed to get IB devices list: Function not implemented

Signed-off-by: Bart Van Assche <bart.vanassche@sandisk.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>