Improve linker flag detection to work with clang and -Werror
If a linker flag is passed to a clang compile command then it
prints out a warning that the flag is a linker flag. If -Werror
is set then the warning is an error and all linker flag detection
is broken.

The fix is to only pass linker flags to the linker. cmake does not have
an integrated way to do this until 3.2, the recommended approach is to
abuse CMAKE_REQUIRED_LIBRARIES on older versions.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>