ibverbs: increase value[] len to read all node_type values
This commit increases the length of the value[] array to 16 because 8
bytes is not long enough to read all valid
/sys/class/infiniband/*/node_type files.  For example:

$ cat /sys/class/infiniband/usnic_0/node_type
6: usNIC UDP

Including \n, this is 13 characters.

More speicifcally, in the kernel IB core,
drivers/inifiniband/core/sysfs.c returns "%d: usNIC UDP\n" as the
value from catenating the node_type file.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>