mlx5: Add support for efficient UAR allocation
The hardware allows now to use UARs with a fixed size of 4KB. To enable
this feature we need both firmware and driver to support this feature in
order to make benefit of this. Benefit is obtained when the system page
size is larger than 4KB, like power pc.

In order to support compatbility between any version of
library/driver/firmware we add a flags field to struct
mlx5_alloc_ucontext to notify the kernel driver whether this is a new
library that supports 4KB UARs. In response, the kernel driver will
return log_uar_size and number of UARs in a page.

We also make order in the terms used. So a "page" now refers to a system
page, "uar" refers to the 4KB UAR area and "uuar" or micro uar refers to
the section in the device registers that is used for posting doorbells.

In case of larger page sizes, we will try to assign more lockless uuars
and preserve the number of medium latency uuars.

Signed-off-by: Eli Cohen <eli@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>