Add -DIN_PLACE=1 to cmake
This configures the paths to run the software from the local build
directory. Much of this is done automatically by how cmake uses rpath,
but the two cases where we dlopen things have to be handled by us.

The simple solution is to hard wire the build directory path into the
binaries.

'etc' is also relocated to build/etc/ as that is part of how the verbs
providers are located.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Tested-by: Steve Wise <swise@opengridcomputing.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>