Provide cmake support to enable C11 stdatomic.h
Turn on C11 support in the compiler, for newer compilers this brings
in the necessary semantics. This flag started working as of gcc 4.7
(released 2012)

gcc 4.7 and 4.8 do not have complete C11 support and requires compat for
stdatomic.h, use a ported FreeBSD header (courtesy of
https://gist.github.com/nhatminhle/5181506) to provide enough stdatomic.h
support on some older compilers.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>