Fix shared library filenames
We expect the filename to reflect the symbol versions in the file,
this means, eg libibverbs1.3 should have the IBVERBS_1.3 as the newest
symbol. The same applies to all libraries.

verbs was not updated in commit 3ca7a103148 when new symbols were added

umad/rdmacm probably legitimately had a '.1' someday, but failed to
follow symver convention. Bump it back down.

Since distros are not using it in packaging the shared library version
and filename is no longer used by anything. We simply want it to follow
our own internal conventions.

Fixes: 3ca7a1031486 ("ibverbs: Add support for packet pacing")
Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>