Make the provider facing libibverbs interface private
The first part is to remove the header files that define the
prototypes for these symbols from the set of packaged headers.
This ensures that nothing can compile and use these symbols.

Next we move the symbols into a new symbol version stanza only
for the private ABI. This breaks every existing out of-tree provider,
but the earlier change to ibv_cmd_create_ah already did that.

There are a few symbols that are still private by virtue of not being
in public headers, but these are used internally by the other libraries.
For distribution sanity continue to treat them as public ABI.

The libibverbs filename version is set back to 1.1 as that reflects
the last public ABI change and travis's check is updated to handle
PRIVATE symbols.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>