debian: Remove duplicate section fields
In debian/control, the section field for a binary package duplicates the
value inherited from the source package paragraph. This doesn't hurt
anything, but you may want to take advantage of the inheritance and set
the value in only one place. It prevents missing duplicate places that
need to be fixed if the value ever changes.

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>