debian: Loosen libibverbs1 dependency
ibverbs-providers uses the private symbols of libibverbs1. Thus
libibverbs1 will not work with an old version version of
ibverbs-providers if the private symbol version is bumped. Instead
requiring installing the same version of ibverbs-providers and
libibverbs1, just declare that libibverbs1 breaks all ibverbs-providers
package versions that use an older private symbol version.

When bumping the private symbol version of libibverbs, do not forget to
bump the version in the Breaks statement.

Signed-off-by: Benjamin Drung <benjamin.drung@profitbricks.com>