mlx4: Add support for WQ control path related verbs
The base IB object to enable RSS functionality is a WQ (i.e. ibv_wq).
This patch implements the related WQ verbs as of create, modify and
destroy.
In downstream patches the WQ will be used as part of an indirection
table to enable RSS QP creation.

Since ConnectX-3 hardware requires consecutive WQNs list as receive
descriptor queues for the RSS QP, a range size for a consecutive
WQNs set is delivered in the driver data of the create WQ related
verb. This range size value will be set by a direct verb which will
be introduced in the downstream patches.

Signed-off-by: Guy Levi <guyle@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>