ibverbs: Add tunnel VXLAN flow specification
Add flow specification for receiving a specific VXLAN packet. The
specification defines a filter on the VNI field in the packet. There is an
option to mask a range of tunnel IDs to be steered by using the filter
mask.

VNI: 24-bit value that provides a unique identifier for the
individual VXLAN segment.

Signed-off-by: Moses Reuben <mosesr@mellanox.com>
Signed-off-by: Artemy Kovalyov <artemyko@mellanox.com>
Reviewed-by: Maor Gottlieb <maorg@mellanox.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>