verbs: Add mmio_wc_spinlock barrier
For x86 the serialization within the spin lock is enough to
strongly order WC and other memory types.

Add a new barrier named 'mmio_wc_spinlock' to optimize
that.

The mmio_wc_spinunlock barrier was added as well, it may be
optimized in some further step based on some required
justification.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>
Reviewed-by: Yishai Hadas <yishaih@mellanox.com>