Patch more headers for sparse usage
Turns out we need quite a bit more than just a simple wrapper around
endian.h. Use a patching scheme instead, we store diffs we want to apply
and a script directly modifies the system glibc headers.

The shim headers are designed to compile, not to work. When building with sparse
the resulting binaries are garbage.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>