verbs: Provide function prototypes for all of the symbol versions
It is a bit goofy, but gcc requires a unique global name for the
implementation when aliasing to the symbol version. So we need to have
two prototypes for the same entry point, one with the __ to avoid the
warning.

This is motivated by sparse which has no way to disable this check.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>