mlx5: Expose direct create_cq with private data
This patch exposes an API named mlx5dv_create_cq to enable creating a
CQ with some mlx5 private data.

Specifically, it enables creating a CQ in a mode that
few CQEs may be compressed into a single CQE.

To use the shared code of libibverbs when a CQ is created it uses the
verbs_init_cq() helper API that was introduced in previous patch.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>