verbs: Expose verbs init CQ helper API
This patch exposes to providers an helper function to init a CQ
post its creation.

The internal library code was refactored to use this helper
API to have a shared code for all creation flows (i.e. create_cq and
create_cq_ex).

As part of this refactoring reduced the scope of the lock that was taken
as part of creating/destroying a CQ to protect only what it was intended
to do. (i.e. channel->refcnt)

It's the application responsibility not to call concurrently to create a
CQ with a given completion channel and in the same time destroy it from
other thread, no change at that point from the original code.

The helper function going to be used in downstream patch in this
series.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>