cmake: Update -DKERNEL_DIR to work with latest kernels
We need to symlink the new ioctl files to be able to use ib_user_mad.h
as well. Since we do not use the new stuff, no copy under buildlib is
provided.

Update this scheme a little bit to better handle all the conditions,
if a required kernel header is missing this now fails at cmake time,
otherwise it is silent for missing headers that are only neeed to make
-DKERNEL_DIR work.

Signed-off-by: Jason Gunthorpe <jgunthorpe@obsidianresearch.com>